requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlmodel==0.0.19
sqlalchemy==2.0.30
chromadb==0.5.0
pydantic>=2.10.0
pydantic-settings==2.3.1
passlib[bcrypt]==1.7.4
argon2-cffi
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
fastapi-mail==1.4.1
redis[hiredis]==5.0.4
alembic==1.13.1
python-multipart==0.0.9
python-dotenv==1.0.1
langchain>=0.2.1
langchain-community>=0.2.1
langchain-core>=0.2.20
langchain-text-splitters>=0.2.0
langchain-google-genai>=1.0.5
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.3
sentence-transformers==2.7.0
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.27.0
googlesearch-python==1.2.3
yfinance==0.2.40
google-generativeai==0.5.4
huggingface-hub==0.23.2
psycopg2-binary
asyncpg
aiomysql
